Address linter remarks
